+++ R/RcppExports.R
@@ -9,3 +9,7 @@ markovChains <- function(nchains, P, chainLen, state0) {
     .Call(`_simstudy_markovChains`, nchains, P, chainLen, state0)
 }
 
+clipVec <- function(id, seq, event) {
+    .Call(`_simstudy_clipVec`, id, seq, event)
+}
+