+++ NEWS.md
@@ -95,7 +95,7 @@
 # simstudy 0.1.14
 
 * Added functions genMarkov and addMarkov to create data.table with (or add to existing data.table) individual chains of Markov processes.
-* Added function genNthEvent to create data.table with binary event outcome in a longintudinal setting. 
+* Added function genNthEvent to create data.table with binary event outcome in a longitudinal setting. 
 * Updated function genCluster so that cluster size can be specified as an integer, and will be constant across all clusters.
 * Updated function addPeriods that period name can be specified.
 * Updated function trtStepWedge so that a transition period can be included.